Move Features class into separate file features.h
