+++ CMakeLists.txt
@@ -25,6 +25,20 @@ set(YARAMOD_SRC_BIN_DIR "${PROJECT_BINARY_DIR}/src")
 set(YARAMOD_DEPS_DIR    "${PROJECT_SOURCE_DIR}/deps")
 set(YARAMOD_PUBLIC_MODULES_DIR     "${PROJECT_SOURCE_DIR}/src/modules")
 set(YARAMOD_PRIVATE_MODULES_DIR     "${PROJECT_SOURCE_DIR}/src/modules/private")
+set(WRAP_MODULE_SOURCES
+	${YARAMOD_SRC_DIR}/modules/module_elf.cpp
+	${YARAMOD_SRC_DIR}/modules/module_macho.cpp
+	${YARAMOD_SRC_DIR}/modules/module_cuckoo.cpp
+	${YARAMOD_SRC_DIR}/modules/module_magic.cpp
+	${YARAMOD_SRC_DIR}/modules/module_dotnet.cpp
+	${YARAMOD_SRC_DIR}/modules/module_pe.cpp
+	${YARAMOD_SRC_DIR}/modules/module_metadata.cpp
+	${YARAMOD_SRC_DIR}/modules/module_dex.cpp
+	${YARAMOD_SRC_DIR}/modules/module_cuckoo_deprecated.cpp
+	${YARAMOD_SRC_DIR}/modules/module_math.cpp
+	${YARAMOD_SRC_DIR}/modules/module_time.cpp
+	${YARAMOD_SRC_DIR}/modules/module_hash.cpp
+)
 
 configure_file("${PROJECT_SOURCE_DIR}/modules_path.h.in" "${PROJECT_SOURCE_DIR}/include/yaramod/modules_path.h")
 
@@ -51,34 +65,17 @@ if(YARAMOD_DOCS)
 	add_subdirectory(docs)
 endif()
 
-set(WRAP_MODULE_SOURCES
-	${YARAMOD_SRC_DIR}/modules/module_elf.cpp
-	${YARAMOD_SRC_DIR}/modules/module_macho.cpp
-	${YARAMOD_SRC_DIR}/modules/module_cuckoo.cpp
-	${YARAMOD_SRC_DIR}/modules/module_magic.cpp
-	${YARAMOD_SRC_DIR}/modules/module_dotnet.cpp
-	${YARAMOD_SRC_DIR}/modules/module_pe.cpp
-	${YARAMOD_SRC_DIR}/modules/module_metadata.cpp
-	${YARAMOD_SRC_DIR}/modules/module_dex.cpp
-	${YARAMOD_SRC_DIR}/modules/module_cuckoo_deprecated.cpp
-	${YARAMOD_SRC_DIR}/modules/module_math.cpp
-	${YARAMOD_SRC_DIR}/modules/module_time.cpp
-	${YARAMOD_SRC_DIR}/modules/module_hash.cpp
-)
-
+# Command that creates modules stated in ${WRAP_MODULE_SOURCES}
 add_custom_command(
 	OUTPUT ${WRAP_MODULE_SOURCES}
+	# Create .cpp modules from .json modules
 	COMMAND	python ${YARAMOD_SRC_DIR}/json_to_byte_array.py -i "${YARAMOD_PUBLIC_MODULES_DIR}/public" -o "${YARAMOD_PUBLIC_MODULES_DIR}"
-	# WORKING_DIRECTORY
-	# 	${CMAKE_CURRENT_SOURCE_DIR}
 	COMMENT
 		"Converting yaramod modules from JSON to byte arrays in .cpp files"
-	# VERBATIM
 )
 
+# Target depending on creation of modules ${WRAP_MODULE_SOURCES} by the command 
 add_custom_target(modules_wrapping ALL
-	# WORKING_DIRECTORY
-	# 	${CMAKE_CURRENT_BINARY_DIR}
 	DEPENDS
 		${WRAP_MODULE_SOURCES}
 	VERBATIM
